+++ src/compas/robots/model/base.py
@@ -90,7 +90,8 @@ class FrameProxy(ProxyObject):
         Examples
         --------
         >>> attributes = {'rpy': '0.0 1.57 0.0', 'xyz': '0.0 0.13 0.0'}
-        >>> FrameProxy.from_urdf(attributes, [], None)
+        >>> f = FrameProxy.from_urdf(attributes, [], None)
+        >>> print(f)
         Frame(Point(0.000, 0.130, 0.000), Vector(0.001, 0.000, -1.000), Vector(0.000, 1.000, 0.000))
 
         """